moved quick action code out of app delegate
